Revert commits adding motion text
As per #9, the motion text we were using was capturing what was said
immediately *after* a vote instead of before it. Therefore the text
wasn't useful and was often out of context.

So far we haven't been able to find useful, reliable text to use. It's
unlikely that we'll be able to unless we get structured debate data
from the parliament. For the moment it's better not to be smart and
just not set the motion text.

Reverts:

096489b Set motion ID to date/time in lieu of an ID from parliament. Fixes #7
60d357a Differentiate variable name
d8ca008 Don't include the timestamp in the motion text
54597c4 Use speech from transcript that matches time of vote as motion_text